janium_functions/read_email_inbox/read_email_inbox_director/requirements.txt
sqlalchemy
pymysql
google-cloud-pubsub
holidays==0.10.3
colorama